+++ demo/views/layout/row-demo/row-demo.js
@@ -9,6 +9,7 @@ import Icon from 'components/icon';
 import Row from 'components/row';
 import Button from 'components/button';
 import Textbox from 'components/textbox';
+import Checkbox from 'components/checkbox';
 import Number from 'components/number';
 import Dropdown from 'components/dropdown';
 import { Table, TableRow, TableCell, TableHeader } from 'components/table';
@@ -34,7 +35,7 @@ class RowDemo extends React.Component {
    */
   get demo() {
     return (
-      <Row columns={ this.value('columns') } gutter={ this.value('gutter') }>
+      <Row columns={ this.value('columns') } gutter={ this.value('gutter') } columnDivide={ this.value('columnDivide') }>
         {
           this.value('columnData').map((data, index) => {
             return (
@@ -66,6 +67,10 @@ class RowDemo extends React.Component {
 
     if (this.value('gutter')) {
       html += `\n  gutter='${this.value('gutter')}'\n`;
+    }
+
+    if (this.value('columnDivide')) {
+      html += `\n  columnDivide='true'\n`;
     }
 
     html += ">\n";
@@ -230,6 +235,12 @@ class RowDemo extends React.Component {
             labelInline={ true }
             onChange={ this.action.bind(this, 'gutter') }
             options={ this.sizeOptions }
+          />
+
+          <Checkbox
+            label="Column Divide"
+            value={ this.value('columnDivide') }
+            onChange={ this.action.bind(this, 'columnDivide') }
           />
         </Row>