+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * `Link` component now has a prop of `iconAlign` to align icons to the right of the link's text.
 * `Row` component can now be given a size to control the size of the gutter using the prop `gutter` (eg. `extra-small`, `small`, `medium`, `large` or `extra-large`).
+* `Row` can enable `columnDivide` to add dividing lines between columns.
 
 ## Minor Improvements